Wordle Game
A simple version of Wordle game. It can be played on terminal. Player has to start the game with a word. Then the program checks the word's letters. If the letter is correct and also it's on correct place, the letter turns green. If the letter is correct but the place is not correct, it turns yellow. If the letter is not correct then it turns red. It goes on until the player finds the correct word or the player's chances finished.